+++ modules/Multiplane/themes/rljbase/views/layouts/tags.php
@@ -1,8 +1,17 @@
-
 <main id="main">
-    @render('views:partials/search-extended.php')
+    <p class="tags">
+        @if(!empty($tags))
+          @lang('All tags:')
+          @foreach($tags as $tag)
+          <a href="@base('/tags/'.urlencode($tag))" class="label tag">{{ $tag }}</a>
+          @endforeach
+        @else
+          @lang('No tags found')
+        @endif
+    </p>
+  @if($count)
     @if (isset($error))<p>@lang($error)</p>@endif
-    <p>{{ count($list) }} @lang($count == 1 ? 'result' : 'results')</p>
+    <p>{{ $count }} @lang($count == 1 ? 'result' : 'results')</p>
     @foreach($list as $l)
     <div class="search-entries" data-weight="{{ $l['weight'] }}">
         <h3><a href="{{ $l['url'] }}">{{ $l['title'] }}</a></h3>
@@ -12,4 +21,5 @@
       @endif
     </div>
     @endforeach
+  @endif
 </main>