+++ modules/Multiplane/themes/rljbase/views/partials/nav-mobile-subnav.php
@@ -2,7 +2,7 @@
               @foreach($nav as $k => $n){% $hasChildren = !empty($n['children']) %}
                 <li class="{{ $n['class'] ?? '' }}{{ $onlyMobile ? 'nav-visible-tiny' : '' }}{{ $onlyMobile && $k == 0 ? ' nav-spacer' : ''}}">
                   @if($hasChildren){% $_id = uniqid('mp-nav-') %}
-                    <input type="checkbox" id="{{ $_id }}" tabindex="-1" /><label for="{{ $_id }}"></label>
+                    <input type="checkbox" id="{{ $_id }}" tabindex="-1" aria-label="@lang('Expand sub menu')" /><label for="{{ $_id }}"></label>
                   @endif
                   @if(isset($n['url']))
                     <a class="{{ $n['active'] ? 'active' : '' }}{{ $hasChildren ? ' dropdown' : '' }}" href="{{ $n['url'] }}">{{{ $n['title'] }}}</a>