+++ modules/Multiplane/themes/rljbase/views/partials/nav-mobile.php
@@ -7,9 +7,9 @@ if (empty($navs)) return;
 ?>
 
         <nav id="nav" class="{{ $class ?? 'horizontal' }}" aria-label="@lang(ucfirst($types[0]))">
-            <a href="#nav" class="icon-menu"></a>
-            <a class="icon-close" href="#" title="@lang('Close')" aria-label="@lang('Close')"></a>
-            <a href="@base('/')" class="icon-home nav-visible-tiny" title="@lang('Home')" aria-label="@lang('Home')"></a>
+            <a href="#nav" class="icon-menu" title="@lang('Expand menu')"></a>
+            <a class="icon-close" href="#" title="@lang('Hide menu')"></a>
+            <a href="{{ mp()->base('/') }}" class="icon-home nav-visible-tiny" title="@lang('Home')"></a>
             <ul>
           @foreach($navs as $k => $nav)
            @if(!empty($nav))