+++ modules/Multiplane/themes/rljbase/views/partials/search-extended.php
@@ -8,8 +8,8 @@ if (!isset($id)) $id = 'search-extended';
             <div class="search-extended">
                 <form action="@base('/search')">
                     <div>
-                        <label for="{{ $id }}">@lang('Search')</label>
-                        <input id="{{ $id }}" name="search" type="text" value="{{{ $search }}}" minlength="{{ mp()->get('search/minLength') }}" aria-label="@lang('Search')" />
+                        <label for="{{ $id }}">@lang('Search in site')</label>
+                        <input id="{{ $id }}" name="search" type="text" value="{{{ $search }}}" minlength="{{ mp()->get('search/minLength') }}" required />
                     </div>
                     <div>
                         <input name="highlight" type="checkbox" value="1" id="mp_search_highlight" {{ $highlight ? 'checked' : '' }} />