+++ modules/Multiplane/themes/rljbase/views/widgets/contactform.php
@@ -1,4 +1,0 @@
-
-<div class="widget contactform">
-{{ $app->helper('form')->form($options['form'] ?? mp()->contact, $options) }} 
-</div>