+++ modules/Multiplane/themes/rljstripes/views/errors/404.php
@@ -3,10 +3,6 @@
             <section>
                 <h1>@lang('Page not found')</h1>
                 <p>@lang("Something went wrong. This site doesn't exist.")</p>
-                <p><a href="@base('/')">@lang('Back to start')</a></p>
-
-            @if(mp()->get('search/enabled'))
-                @render('views:partials/search.php')
-            @endif
+                <p><a href="{{ mp()->base('/') }}">@lang('Back to start')</a></p>
             </section>
         </main>